+++ ruststep-derive/src/select.rs
@@ -16,7 +16,6 @@ struct Input {
     variant_into_exprs: Vec<TokenStream2>,
     holder_types: Vec<syn::Type>,
     holder_exprs: Vec<TokenStream2>,
-    table_fields: Vec<Option<syn::Ident>>,
 }
 
 impl Input {
@@ -36,16 +35,10 @@ impl Input {
 
         let mut holder_exprs = Vec::new();
         let mut holder_types = Vec::new();
-        let mut table_fields = Vec::new();
         let mut variant_exprs = Vec::new();
         let mut variant_into_exprs = Vec::new();
         for var in &e.variants {
-            let HolderAttr {
-                field,
-                place_holder,
-                ..
-            } = HolderAttr::parse(&var.attrs);
-            table_fields.push(field);
+            let HolderAttr { place_holder, .. } = HolderAttr::parse(&var.attrs);
 
             assert_eq!(var.fields.len(), 1);
             for f in &var.fields {
@@ -88,7 +81,6 @@ impl Input {
             variant_into_exprs,
             holder_types,
             holder_exprs,
-            table_fields,
         }
     }
 
@@ -219,7 +211,7 @@ impl Input {
             ident,
             holder_ident,
             variants,
-            table_fields,
+            holder_types,
             table,
             variant_into_exprs,
             ..
@@ -227,16 +219,12 @@ impl Input {
         let ruststep = ruststep_crate();
         let itertools = itertools_crate();
         let mut vars = Vec::new();
-        let mut fields = Vec::new();
         let mut exprs = Vec::new();
-        for ((var, field), expr) in variants
-            .iter()
-            .zip(table_fields.iter())
-            .zip(variant_into_exprs.iter())
-        {
-            if let Some(field) = field {
+        let mut holders = Vec::<syn::Type>::new();
+        for ((var, holder), expr) in variants.iter().zip(holder_types).zip(variant_into_exprs) {
+            if let FieldType::Boxed(path) = holder.clone().try_into().unwrap() {
                 vars.push(var);
-                fields.push(field);
+                holders.push(path.as_ref().clone().into());
                 exprs.push(expr);
             }
         }
@@ -245,7 +233,7 @@ impl Input {
             impl #ruststep::tables::EntityTable<#holder_ident> for #table {
                 fn get_owned(&self, entity_id: u64) -> #ruststep::error::Result<#ident> {
                     #(
-                    if let Ok(owned) = #ruststep::tables::get_owned(self, &self.#fields, entity_id) {
+                    if let Ok(owned) = #ruststep::tables::EntityTable::<#holders>::get_owned(self, entity_id) {
                         return Ok(#ident::#vars(#exprs));
                     }
                     )*
@@ -254,7 +242,7 @@ impl Input {
                 fn owned_iter<'table>(&'table self) -> Box<dyn Iterator<Item = #ruststep::error::Result<#ident>> + 'table> {
                     Box::new(#itertools::chain![
                         #(
-                        #ruststep::tables::owned_iter(self, &self.#fields)
+                        #ruststep::tables::EntityTable::<#holders>::owned_iter(self)
                             .map(|owned| owned.map(|owned| #ident::#vars(#exprs)))
                         ),*
                     ])